+++ alphadia/search_plan.py
@@ -12,12 +12,12 @@ from alphadia.constants.keys import ConfigKeys, StatOutputKeys
 from alphadia.outputtransform.search_plan_output import (
     SearchPlanOutput,
 )
+from alphadia.reporting import reporting
 from alphadia.search_step import (
     SearchStep,
     logger,
 )
 from alphadia.utilities.logging import print_environment, print_logo
-from alphadia.workflow import reporting
 
 # TODO the names of the steps need to be adjusted
 TRANSFER_STEP_NAME = "transfer"
