+++ tests/unit_tests/test_workflow.py
@@ -15,7 +15,8 @@ from sklearn.linear_model import LinearRegression
 from alphadia.calibration.models import LOESSRegression
 from alphadia.calibration.property import Calibration
 from alphadia.fdrexperimental import BinaryClassifierLegacyNewBatching
-from alphadia.workflow import base, manager, optimization, peptidecentric, reporting
+from alphadia.reporting import reporting
+from alphadia.workflow import base, manager, optimization, peptidecentric
 from alphadia.workflow.config import Config
